Add validation to decrease unneccessary requests
This validate the secret key to some point at least an empty string and
must be at least 10 min length.

Validate the request as well.

The idea is to decrease the number of unneccessary request to the API as
some basic checks can be made on the users' server before making request
to paystack server.